LayoutTests/platform/win-wk2/platform/win/plugins/window-region-is-set-to-clip-rect-expected.txt

CONSOLE MESSAGE: line 0: PLUGIN: ::GetWindowRgnBox failed with error 0

This tests that a plugin's window's window region is set to its clip rect.
